feat: Integrate Gutenberg block data
This commit adds a few integration points with the new Gutenberg editor,
mainly by parsing blocks when a post saves and storing the block data
as post meta in serialized form. Then, the plugin exposes this block data
via the REST API in posts with a custom property called `headlessBlocks`.